requirements.txt
streamlit==1.41.1
opencv-python-headless==4.11.0.86
numpy>=1.26.0
Pillow>=10.0.0
pandas>=2.1.0
python-docx>=1.1.0
reportlab>=4.1.0
openpyxl>=3.1.0
